moved primitives to separate lib
