+++ primitives/hash.h
